+++ Batch_1_Python_Programs/print_the_bigger_num.py
@@ -1,5 +1,5 @@
-num_1 = int(input("enter your first number: "))
-num_2 = int(input("enter your second number: "))
+num_1 = float(input("enter your first number: "))
+num_2 = float(input("enter your second number: "))
 if num_1>num_2:
     print(num_1)
 elif num_2>num_1:
